backend/pyproject.toml
[project]
name = "fastapi-ncp-mailing"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dongyeonyug", email = "yugdongyeon@gmail.com"},
]
dependencies = ["fastapi>=0.127.0", "uvicorn>=0.40.0", "sqlalchemy>=2.0.45", "passlib[bcrypt]>=1.7.4", "bcrypt==4.0.1", "python-jose>=3.5.0", "redis>=7.1.0", "jinja2>=3.1.6"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
